refactored gitlab provider
