+++ src/SimpSpector/Provider/Event/EventInterface.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace SimpSpector\Provider\Event;
+
+/**
+ * @author Simon Mönch <simonmoench@gmail.com>
+ */
+interface EventInterface
+{
+
+}
