+++ src/SimpleThings/AppBundle/Resources/config/gitlab.xml
@@ -6,24 +6,23 @@
 
     <services>
 
-        <service id="simple_things_app.gitlab.commit_factory" class="SimpleThings\AppBundle\GitLab\CommitFactory">
-            <argument type="service" id="doctrine.orm.entity_manager"/>
-            <argument type="service" id="gitlab_api"/>
-        </service>
-
-        <service id="simple_things_app.gitlab.request_handler" class="SimpleThings\AppBundle\GitLab\RequestHandler">
+        <service id="simple_things_app.gitlab.request_handler" class="SimpSpector\Provider\RequestHandler">
             <argument type="service" id="doctrine.orm.default_entity_manager"/>
             <argument type="service" id="simpspector.app.repository.merge_request"/>
             <argument type="service" id="simpspector.app.repository.project"/>
-            <argument type="service" id="gitlab_api"/>
+            <argument type="service" id="simpspector.provider.gitlab_adapter"/>
             <argument type="service" id="simple_things_app.gitlab.notifier"/>
             <argument type="service" id="logger"/>
         </service>
 
-        <service id="simple_things_app.gitlab.notifier" class="SimpleThings\AppBundle\GitLab\Notifier">
+        <service id="simple_things_app.gitlab.notifier" class="SimpSpector\Provider\Notifier">
+            <argument type="service" id="simpspector.provider.gitlab_adapter"/>
+            <argument type="service" id="logger"/>
+        </service>
+
+        <service id="simpspector.provider.gitlab_adapter" class="SimpSpector\Provider\GitlabProviderAdapter">
             <argument type="service" id="gitlab_api"/>
             <argument type="service" id="simple_things_app.badge_generator"/>
-            <argument type="service" id="logger"/>
         </service>
 
     </services>
